Changed some control binding, made 2 control teleOp. Cleaned up the Auto OpModes
