+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/Controls/ControlOperator.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.twenty403.Controls;
+
+public class ControlOperator {
+}
