+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/Controls/Controls2.java
@@ -1,6 +1,6 @@
-package org.firstinspires.ftc.twenty403;
+package org.firstinspires.ftc.twenty403.Controls;
 
-import org.firstinspires.ftc.twenty403.Robot.RobotConstant;
+import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.claw.ClawCloseCommand;
 import org.firstinspires.ftc.twenty403.command.claw.ClawOpenCommand;
 import org.firstinspires.ftc.twenty403.command.drive.DriveCommand;
@@ -21,71 +21,69 @@ import org.firstinspires.ftc.twenty403.helpers.BothButtons;
 
 import com.technototes.library.command.CommandScheduler;
 import com.technototes.library.command.ConditionalCommand;
+import com.technototes.library.control.CommandAxis;
 import com.technototes.library.control.CommandButton;
 import com.technototes.library.control.CommandGamepad;
 import com.technototes.library.control.Stick;
 
-public class Controls {
-
+public class Controls2 {
     public Robot robot;
     public CommandGamepad gamepad;
 
     public Stick driveLeftStick, driveRightStick;
-    public CommandButton resetGyroButton, driveStraightenButton, liftUpButton, clawOpenButton;
-    public CommandButton liftDownButton, clawCloseButton, liftIntakePos;
-    public CommandButton liftGroundOrOverrideDown, liftLow, liftMedium, liftHighOrOverrideUp, liftOverrideZeroButton;
-    public CommandButton turboButton;
-
+    public CommandButton resetGyroButton, driveStraighten, turboButton;
+    public CommandButton liftDownButton, liftUpButton, clawOpenButton, clawCloseButton, liftOverrideZeroButton;
+    public CommandButton liftGroundOrOverrideDown, liftLowOrOverrideUp, liftMedium, liftHigh, liftIntakePos;
+    public CommandAxis driveStraight;
     public BothButtons override;
 
-    public Controls(CommandGamepad g, Robot r) {
+    public Controls2(CommandGamepad g, Robot r) {
         this.robot = r;
         gamepad = g;
-
-        override = new BothButtons(g.leftTrigger.getAsButton(0.5), g.rightTrigger.getAsButton(.5));
+        override = new BothButtons(g.leftTrigger.getAsButton(0.5));
 
         AssignNamedControllerButton();
 
-        if (RobotConstant.DRIVE_CONNECTED) {
+        if (Robot.RobotConstant.DRIVE_CONNECTED) {
             bindDriveControls();
         }
-        if (RobotConstant.CLAW_CONNECTED) {
+        if (Robot.RobotConstant.CLAW_CONNECTED) {
             bindClawControls();
         }
-        if (RobotConstant.LIFT_CONNECTED) {
+        if (Robot.RobotConstant.LIFT_CONNECTED) {
             bindLiftControls();
         }
     }
 
     private void AssignNamedControllerButton() {
         resetGyroButton = gamepad.rightStickButton;
-        liftOverrideZeroButton = gamepad.share;
         driveLeftStick = gamepad.leftStick;
         driveRightStick = gamepad.rightStick;
-        turboButton = gamepad.triangle;
-        liftUpButton = gamepad.square;
-        liftDownButton = gamepad.cross;
+        turboButton = gamepad.leftStickButton;
+        driveStraight = gamepad.rightTrigger;
 
-        liftIntakePos = gamepad.circle;
+        liftUpButton = gamepad.dpadRight;
+        liftDownButton = gamepad.dpadDown;
+        liftIntakePos = gamepad.dpadLeft;
 
-        clawOpenButton = gamepad.leftBumper;
+        clawOpenButton = gamepad.rightBumper;
+        clawCloseButton = gamepad.leftBumper;
 
-        clawCloseButton = gamepad.leftTrigger.getAsButton();
+        liftMedium = gamepad.circle;
+        liftHigh = gamepad.triangle;
 
-        liftGroundOrOverrideDown = gamepad.dpadDown;
+        liftGroundOrOverrideDown = gamepad.cross;
+        liftLowOrOverrideUp = gamepad.square;
+        liftOverrideZeroButton = gamepad.triangle;
 
-        liftLow = gamepad.dpadLeft;
-        liftMedium = gamepad.dpadRight;
-        liftHighOrOverrideUp = gamepad.dpadUp;
 
-        // TODO: Identify other controls for drive straighten button
-        driveStraightenButton = gamepad.options;
+        // TODO: Identify other controls for
     }
 
     public void bindDriveControls() {
         CommandScheduler.getInstance()
-                .scheduleJoystick(new DriveCommand(
-                        robot.drivebaseSubsystem, driveLeftStick, driveRightStick, driveStraightenButton));
+                .scheduleJoystick(
+                        new DriveCommand(robot.drivebaseSubsystem, driveLeftStick, driveRightStick, driveStraight));
         turboButton.whenPressed(new TurboCommand(robot.drivebaseSubsystem));
         turboButton.whenReleased(new SlowCommand(robot.drivebaseSubsystem));
         // TODO: We probably want buttons to reset the Gyro...
@@ -95,7 +93,7 @@ public class Controls {
     public void bindClawControls() {
         // TODO: Name & Bind claw controls
         clawOpenButton.whenPressed(new ClawOpenCommand(robot.clawSubsystem));
-        clawCloseButton.whenPressed(new ClawCloseCommand(robot.clawSubsystem));
+        clawCloseButton.whenReleased(new ClawCloseCommand(robot.clawSubsystem));
     }
 
     public void bindLiftControls() {
@@ -103,18 +101,18 @@ public class Controls {
         liftUpButton.whenPressed(new LiftUpCommand(robot.liftSubsystem));
         liftDownButton.whenPressed(new LiftDownCommand(robot.liftSubsystem));
         liftIntakePos.whenPressed(new LiftIntakeCommand(robot.liftSubsystem));
+        liftOverrideZeroButton.whenPressed(
+                new ConditionalCommand(override, new LiftSetZeroCommand(robot.liftSubsystem)));
 
-        liftLow.whenPressed(new LiftLowJunctionCommand(robot.liftSubsystem));
         liftGroundOrOverrideDown.whenPressed(new ConditionalCommand(
                 override,
                 new LiftMoveDownOverrideCommand(robot.liftSubsystem),
                 new LiftGroundJunctionCommand(robot.liftSubsystem)));
-        liftMedium.whenPressed(new LiftMidJunctionCommand(robot.liftSubsystem));
-        liftHighOrOverrideUp.whenPressed(new ConditionalCommand(
+        liftLowOrOverrideUp.whenPressed(new ConditionalCommand(
                 override,
                 new LiftMoveUpOverrideCommand(robot.liftSubsystem),
-                new LiftHighJunctionCommand(robot.liftSubsystem)));
-        liftOverrideZeroButton.whenPressed(
-                new ConditionalCommand(override, new LiftSetZeroCommand(robot.liftSubsystem)));
+                new LiftLowJunctionCommand(robot.liftSubsystem)));
+        liftMedium.whenPressed(new LiftMidJunctionCommand(robot.liftSubsystem));
+        liftHigh.whenPressed(new LiftHighJunctionCommand(robot.liftSubsystem));
     }
 }
