+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/LeftJustPark.java
@@ -1,13 +1,12 @@
 package org.firstinspires.ftc.twenty403.opmode;
 
-import org.firstinspires.ftc.twenty403.Controls;
+import org.firstinspires.ftc.twenty403.Controls.Controls;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.VisionCommand;
 import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsBlue;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
-import org.firstinspires.ftc.twenty403.command.autonomous.blue_home.AutoBlueHomeParkingSelectionPreLoadCommand;
-import org.firstinspires.ftc.twenty403.command.claw.ClawCloseCommand;
+import org.firstinspires.ftc.twenty403.command.autonomous.blue_away.AutoBlueAwayParkingSelectionJustParkCommand;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
@@ -16,8 +15,8 @@ import com.technototes.library.command.SequentialCommandGroup;
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
-@Autonomous(name = "BlueHomePreLoad")
-public class BlueHomePreLoad extends CommandOpMode {
+@Autonomous(name = "Left Just Park")
+public class LeftJustPark extends CommandOpMode {
     public Robot robot;
     public Controls controls;
     public Hardware hardware;
@@ -25,17 +24,13 @@ public class BlueHomePreLoad extends CommandOpMode {
     @Override
     public void uponInit() {
         hardware = new Hardware(hardwareMap);
-        robot = new Robot(hardware, Alliance.BLUE, StartingPosition.HOME);
-        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsBlue.Home.START.toPose());
+        robot = new Robot(hardware, Alliance.BLUE, StartingPosition.AWAY);
+        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsBlue.Away.START.toPose());
         CommandScheduler.getInstance()
                 .scheduleForState(
                         new SequentialCommandGroup(
-                                new ClawCloseCommand(robot.clawSubsystem),
-                                new AutoBlueHomeParkingSelectionPreLoadCommand(
-                                        robot.visionSystem,
-                                        robot.drivebaseSubsystem,
-                                        robot.clawSubsystem,
-                                        robot.liftSubsystem),
+                                new AutoBlueAwayParkingSelectionJustParkCommand(
+                                        robot.drivebaseSubsystem, robot.visionSystem),
                                 CommandScheduler.getInstance()::terminateOpMode),
                         CommandOpMode.OpModeState.RUN);
         if (Robot.RobotConstant.CAMERA_CONNECTED) {
