+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/LeftPreLoad.java
@@ -1,11 +1,12 @@
 package org.firstinspires.ftc.twenty403.opmode;
 
+import org.firstinspires.ftc.twenty403.Controls.Controls;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.VisionCommand;
-import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsRed;
+import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsBlue;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
-import org.firstinspires.ftc.twenty403.command.autonomous.red_away.AutoRedAwayParkingSelectionPreLoadCommand;
+import org.firstinspires.ftc.twenty403.command.autonomous.blue_away.AutoBlueAwayParkingSelectionPreLoadCommand;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
@@ -14,24 +15,25 @@ import com.technototes.library.command.SequentialCommandGroup;
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
 
-@Autonomous(name = "RedAwayPreLoad")
-public class RedAwayPreload extends CommandOpMode {
+@Autonomous(name = "Right PreLoad")
+public class LeftPreLoad extends CommandOpMode {
     public Robot robot;
+    public Controls controls;
     public Hardware hardware;
 
     @Override
     public void uponInit() {
         hardware = new Hardware(hardwareMap);
-        robot = new Robot(hardware, Alliance.RED, StartingPosition.AWAY);
-        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsRed.Away.START.toPose());
+        robot = new Robot(hardware, Alliance.BLUE, StartingPosition.AWAY);
+        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsBlue.Away.START.toPose());
         CommandScheduler.getInstance()
                 .scheduleForState(
                         new SequentialCommandGroup(
-                                new AutoRedAwayParkingSelectionPreLoadCommand(
-                                        robot.visionSystem,
+                                new AutoBlueAwayParkingSelectionPreLoadCommand(
                                         robot.drivebaseSubsystem,
-                                        robot.clawSubsystem,
-                                        robot.liftSubsystem),
+                                        robot.visionSystem,
+                                        robot.liftSubsystem,
+                                        robot.clawSubsystem),
                                 CommandScheduler.getInstance()::terminateOpMode),
                         CommandOpMode.OpModeState.RUN);
         if (Robot.RobotConstant.CAMERA_CONNECTED) {
