+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/OldAutos/RedHomePark.java
@@ -1,38 +1,37 @@
-package org.firstinspires.ftc.twenty403.opmode;
+package org.firstinspires.ftc.twenty403.opmode.OldAutos;
 
-import org.firstinspires.ftc.twenty403.Controls;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.VisionCommand;
-import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsBlue;
+import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsRed;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
-import org.firstinspires.ftc.twenty403.command.autonomous.blue_away.AutoBlueAwayParkingSelectionJustParkCommand;
+import org.firstinspires.ftc.twenty403.command.autonomous.red_home.AutoRedHomeParkingSelectionJustParkCommand;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.technototes.library.command.CommandScheduler;
 import com.technototes.library.command.SequentialCommandGroup;
 import com.technototes.library.structure.CommandOpMode;
 import com.technototes.library.util.Alliance;
-
-@Autonomous(name = "Blue Away Just Park")
-public class BlueAwayJustPark extends CommandOpMode {
+@Disabled
+@Autonomous(name = "RedHomeParkOnly")
+public class RedHomePark extends CommandOpMode {
     public Robot robot;
-    public Controls controls;
     public Hardware hardware;
 
     @Override
     public void uponInit() {
         hardware = new Hardware(hardwareMap);
-        robot = new Robot(hardware, Alliance.BLUE, StartingPosition.AWAY);
-        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsBlue.Away.START.toPose());
+        robot = new Robot(hardware, Alliance.RED, StartingPosition.HOME);
+        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsRed.Home.START.toPose());
         CommandScheduler.getInstance()
                 .scheduleForState(
                         new SequentialCommandGroup(
-                                new AutoBlueAwayParkingSelectionJustParkCommand(
+                                new AutoRedHomeParkingSelectionJustParkCommand(
                                         robot.drivebaseSubsystem, robot.visionSystem),
                                 CommandScheduler.getInstance()::terminateOpMode),
-                        CommandOpMode.OpModeState.RUN);
+                        OpModeState.RUN);
         if (Robot.RobotConstant.CAMERA_CONNECTED) {
             CommandScheduler.getInstance().scheduleInit(new VisionCommand(robot.visionSystem));
         }
