--- Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/RedHomePreLoad.java
@@ -1,41 +1,0 @@
-package org.firstinspires.ftc.twenty403.opmode;
-
-import org.firstinspires.ftc.twenty403.Hardware;
-import org.firstinspires.ftc.twenty403.Robot;
-import org.firstinspires.ftc.twenty403.command.VisionCommand;
-import org.firstinspires.ftc.twenty403.command.autonomous.AutoConstantsRed;
-import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
-import org.firstinspires.ftc.twenty403.command.autonomous.red_home.AutoRedHomeParkingSelectionPreLoadCommand;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-import com.technototes.library.command.CommandScheduler;
-import com.technototes.library.command.SequentialCommandGroup;
-import com.technototes.library.structure.CommandOpMode;
-import com.technototes.library.util.Alliance;
-
-@Autonomous(name = "RedHomePreLoad")
-public class RedHomePreLoad extends CommandOpMode {
-    public Robot robot;
-    public Hardware hardware;
-
-    @Override
-    public void uponInit() {
-        hardware = new Hardware(hardwareMap);
-        robot = new Robot(hardware, Alliance.RED, StartingPosition.HOME);
-        robot.drivebaseSubsystem.setPoseEstimate(AutoConstantsRed.Home.START.toPose());
-        CommandScheduler.getInstance()
-                .scheduleForState(
-                        new SequentialCommandGroup(
-                                new AutoRedHomeParkingSelectionPreLoadCommand(
-                                        robot.visionSystem,
-                                        robot.drivebaseSubsystem,
-                                        robot.clawSubsystem,
-                                        robot.liftSubsystem),
-                                CommandScheduler.getInstance()::terminateOpMode),
-                        CommandOpMode.OpModeState.RUN);
-        if (Robot.RobotConstant.CAMERA_CONNECTED) {
-            CommandScheduler.getInstance().scheduleInit(new VisionCommand(robot.visionSystem));
-        }
-    }
-}
